backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standart]==0.24.0
python-jose==3.3.0
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0
bleach==6.1.0
python-multipart==0.0.6
pillow==10.1.0
pydantic[email]==2.5.2
passlib==1.7.4
fastapi_mail==1.4.1
asyncpg==0.29.0
pydantic-settings==2.1.0
